refactor(client): extract server tests to sibling file
